fix(interface/api): money is int64 not float64

+++ internal/interface/api/request/payment.go
@@ -1,7 +1,7 @@
 package request
 
 type CreatePayment struct {
-	Amount            float64 `json:"amount" binding:"required"`
+	Amount            int64 `json:"amount" binding:"required"`
 	PaymentProviderID int     `json:"paymentProviderId" binding:"required"`
 	OrderID           int     `json:"orderId" binding:"required"`
 }
