+++ internal/interface/api/response/payment.go
@@ -9,7 +9,7 @@ type (
 
 type Payment struct {
 	ID                int             `json:"id" binding:"required"`
-	Amount            float64         `json:"amount" binding:"required"`
+	Amount            int64         `json:"amount" binding:"required"`
 	PaymentStatus     PaymentStatus   `json:"paymentStatusId" binding:"required"`
 	PaymentProviderID PaymentProvider `json:"paymentProviderId" binding:"omitempty"`
 	UpdatedAt         time.Time       `json:"updatedAt" binding:"required"`
